CC-2049 Email with unsubscription link

+++ Bundles/AvailabilityNotification/src/Spryker/Zed/AvailabilityNotification/Business/Subscription/AvailabilityUnsubscriptionProcessor.php
@@ -9,7 +9,9 @@ namespace Spryker\Zed\AvailabilityNotification\Business\Subscription;
 
 use Generated\Shared\Transfer\AvailabilitySubscriptionResponseTransfer;
 use Generated\Shared\Transfer\AvailabilitySubscriptionTransfer;
+use Spryker\Zed\AvailabilityNotification\Dependency\Facade\AvailabilityNotificationToProductFacadeInterface;
 use Spryker\Zed\AvailabilityNotification\Persistence\AvailabilityNotificationEntityManagerInterface;
+use Spryker\Zed\AvailabilityNotification\Persistence\AvailabilityNotificationRepositoryInterface;
 
 class AvailabilityUnsubscriptionProcessor implements AvailabilityUnsubscriptionProcessorInterface
 {
@@ -19,11 +21,28 @@ class AvailabilityUnsubscriptionProcessor implements AvailabilityUnsubscriptionP
     protected $entityManager;
 
     /**
-     * @param \Spryker\Zed\AvailabilityNotification\Persistence\AvailabilityNotificationEntityManagerInterface $entityManager
+     * @var \Spryker\Zed\AvailabilityNotification\Persistence\AvailabilityNotificationRepositoryInterface
      */
-    public function __construct(AvailabilityNotificationEntityManagerInterface $entityManager)
-    {
+    protected $availabilityNotificationRepository;
+
+    /**
+     * @var \Spryker\Zed\AvailabilityNotification\Dependency\Facade\AvailabilityNotificationToProductFacadeInterface
+     */
+    protected $productFacade;
+
+    /**
+     * @param \Spryker\Zed\AvailabilityNotification\Persistence\AvailabilityNotificationEntityManagerInterface $entityManager
+     * @param \Spryker\Zed\AvailabilityNotification\Persistence\AvailabilityNotificationRepositoryInterface $availabilityNotificationRepository
+     * @param \Spryker\Zed\AvailabilityNotification\Dependency\Facade\AvailabilityNotificationToProductFacadeInterface $productFacade
+     */
+    public function __construct(
+        AvailabilityNotificationEntityManagerInterface $entityManager,
+        AvailabilityNotificationRepositoryInterface $availabilityNotificationRepository,
+        AvailabilityNotificationToProductFacadeInterface $productFacade
+    ) {
         $this->entityManager = $entityManager;
+        $this->availabilityNotificationRepository = $availabilityNotificationRepository;
+        $this->productFacade = $productFacade;
     }
 
     /**
@@ -35,8 +54,19 @@ class AvailabilityUnsubscriptionProcessor implements AvailabilityUnsubscriptionP
     {
         $availabilitySubscriptionTransfer->requireSubscriptionKey();
 
+        $availabilitySubscriptionTransfer = $this->availabilityNotificationRepository->findOneBySubscriptionKey($availabilitySubscriptionTransfer->getSubscriptionKey());
+
+        if ($availabilitySubscriptionTransfer === null) {
+            return (new AvailabilitySubscriptionResponseTransfer())
+                ->setIsSuccess(false)
+                ->setErrorMessage('Subscription doesn\'t exist');
+        }
+
+        $productConcreteTransfer = $this->productFacade->getProductConcrete($availabilitySubscriptionTransfer->getSku());
         $this->entityManager->deleteBySubscriptionKey($availabilitySubscriptionTransfer->getSubscriptionKey());
 
-        return (new AvailabilitySubscriptionResponseTransfer())->setIsSuccess(true);
+        return (new AvailabilitySubscriptionResponseTransfer())
+            ->setIsSuccess(true)
+            ->setProduct($productConcreteTransfer);
     }
 }
